changing ack / data type to int

+++ Project2/src/org/base/Constants.java
@@ -1,7 +1,7 @@
 package org.base;
 
 public class Constants {
-	public static short AckPacket = (short) 43690;
-	public static short DataPacket = 21845;
+	public static int AckPacket =  43690;
+	public static int DataPacket = 21845;
 	public static String GROUP_IP = "203.0.113.0";
 }
